UtilFormatter::formatBytes() added

+++ src/WhyooOs/Util/UtilFormatter.php
@@ -71,7 +71,7 @@ class UtilFormatter
         $base = log($size, 1024);
         $suffixes = ['B', 'KB', 'MB', 'GB', 'TB'];
 
-        return round(pow(1024, $base - floor($base)), $precision) . ' ' . $suffixes[floor($base)];
+        return round(pow(1024, $base - floor($base)), $precision) . ' ' . $suffixes[(int)floor($base)];
     }
